requirements.txt
asgiref==3.11.0
boto3==1.42.10
botocore==1.42.10
certifi==2025.11.12
charset-normalizer==3.4.4
coverage==7.13.0
Django==6.0
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
execnet==2.1.2
factory_boy==3.3.3
Faker==38.2.0
idna==3.11
iniconfig==2.3.0
jmespath==1.0.1
moto==5.0.23
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.4
stripe==14.0.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
